Advance is wrong.

+++ code/systems/text/character.hpp
@@ -13,7 +13,7 @@ struct Character
   float uv[2];
   float st[2];
   uint32_t size[2];
-  uint32_t advance[2];
+  float advance[2];
 };
